requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# AI Services
openai==1.12.0
google-generativeai==0.3.2

# HTTP & API
requests==2.31.0
httpx==0.24.1

# Security
bcrypt==4.1.2
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9

# Image Processing
Pillow==10.2.0

# Utilities
pydantic[email]==2.5.3
